chore(cleanup): properly separate models and interfaces
