+++ README.md
@@ -91,60 +91,60 @@ from causy.graph_model import graph_model_factory
 from causy.common_pipeline_steps.logic import Loop
 from causy.common_pipeline_steps.calculation import CalculatePearsonCorrelations
 from causy.independence_tests.common import (
-  CorrelationCoefficientTest,
-  PartialCorrelationTest,
-  ExtendedPartialCorrelationTestMatrix,
+    CorrelationCoefficientTest,
+    PartialCorrelationTest,
+    ExtendedPartialCorrelationTestMatrix,
 )
 from causy.orientation_rules.pc import (
-  ColliderTest,
-  NonColliderTest,
-  FurtherOrientTripleTest,
-  OrientQuadrupleTest,
-  FurtherOrientQuadrupleTest,
+    ColliderTest,
+    NonColliderTest,
+    FurtherOrientTripleTest,
+    OrientQuadrupleTest,
+    FurtherOrientQuadrupleTest,
 )
-from causy.interfaces import CausyAlgorithm
+from causy.models import CausyAlgorithm
 from causy.algorithms.pc import PC_EDGE_TYPES
 from causy.graph_utils import retrieve_edges
 
 CustomPC = graph_model_factory(
-  CausyAlgorithm(
-    pipeline_steps=[
-      CalculatePearsonCorrelations(),
-      CorrelationCoefficientTest(threshold=0.05),
-      PartialCorrelationTest(threshold=0.05),
-      ExtendedPartialCorrelationTestMatrix(threshold=0.05),
-      ColliderTest(),
-      Loop(
+    CausyAlgorithm(
         pipeline_steps=[
-          NonColliderTest(),
-          FurtherOrientTripleTest(),
-          OrientQuadrupleTest(),
-          FurtherOrientQuadrupleTest(),
+            CalculatePearsonCorrelations(),
+            CorrelationCoefficientTest(threshold=0.05),
+            PartialCorrelationTest(threshold=0.05),
+            ExtendedPartialCorrelationTestMatrix(threshold=0.05),
+            ColliderTest(),
+            Loop(
+                pipeline_steps=[
+                    NonColliderTest(),
+                    FurtherOrientTripleTest(),
+                    OrientQuadrupleTest(),
+                    FurtherOrientQuadrupleTest(),
+                ],
+                exit_condition=ExitOnNoActions(),
+            ),
         ],
-        exit_condition=ExitOnNoActions(),
-      ),
-    ],
-    name="CustomPC",
-    edge_types=PC_EDGE_TYPES,
-  )
+        name="CustomPC",
+        edge_types=PC_EDGE_TYPES,
+    )
 )
 
 model = CustomPC()
 
 model.create_graph_from_data(
-  [
-    {"a": 1, "b": 0.3},
-    {"a": 0.5, "b": 0.2}
-  ]
+    [
+        {"a": 1, "b": 0.3},
+        {"a": 0.5, "b": 0.2}
+    ]
 )
 model.create_all_possible_edges()
 model.execute_pipeline_steps()
 edges = retrieve_edges(model.graph)
 
 for edge in edges:
-  print(
-    f"{edge[0].name} -> {edge[1].name}: {model.graph.edges[edge[0]][edge[1]]}"
-  )
+    print(
+        f"{edge[0].name} -> {edge[1].name}: {model.graph.edges[edge[0]][edge[1]]}"
+    )
 ```
 
 ### Supported algorithms
