+++ causy/workspaces/templates/dataloader.py.tpl
@@ -1,6 +1,6 @@
+from causy.data_loader import AbstractDataLoader
 
-
-class DataLoader:
+class DataLoader(AbstractDataLoader):
     """
     A causy dataloader.
     
@@ -12,7 +12,7 @@ class DataLoader:
         """
         pass
     
-    def load_data(self):
+    def load(self):
         """
         Load the data. This function should yield the data row by row. You have to yield it as a dictionary of column names and values (floats).
         :return: 
@@ -22,7 +22,7 @@ class DataLoader:
             "column2": 2.0
         }
         
-    def data_hash(self):
+    def hash(self):
         """
         Returns a hash of the data. This is useful to check if the data has changed.
         :return:
