+++ tests/test_orientation_rules.py
@@ -1,6 +1,6 @@
 from causy.graph import GraphManager
 from causy.graph_model import graph_model_factory
-from causy.interfaces import TestResult, TestResultAction, CausyAlgorithm
+from causy.models import TestResultAction, TestResult, CausyAlgorithm
 from causy.orientation_rules.fci import ColliderRuleFCI
 
 from tests.utils import CausyTestCase
