+++ tests/test_pc_e2e.py
@@ -1,17 +1,12 @@
-import csv
-import torch
-
-from causy.algorithms import PC, ParallelPC
-from causy.algorithms.pc import PCStable, PC_EDGE_TYPES
+from causy.algorithms.pc import PC_EDGE_TYPES
 from causy.common_pipeline_steps.calculation import CalculatePearsonCorrelations
 from causy.graph_model import graph_model_factory
-from causy.graph_utils import retrieve_edges
 from causy.independence_tests.common import (
     CorrelationCoefficientTest,
     PartialCorrelationTest,
     ExtendedPartialCorrelationTestMatrix,
 )
-from causy.interfaces import CausyAlgorithm
+from causy.models import CausyAlgorithm
 from causy.orientation_rules.pc import ColliderTest
 from causy.sample_generator import IIDSampleGenerator, SampleEdge, NodeReference
