+++ tests/test_serialization.py
@@ -2,7 +2,7 @@ from causy.algorithms.pc import PC_DEFAULT_THRESHOLD
 from causy.common_pipeline_steps.calculation import CalculatePearsonCorrelations
 from causy.graph_model import graph_model_factory
 from causy.independence_tests.common import CorrelationCoefficientTest
-from causy.interfaces import CausyAlgorithm
+from causy.models import CausyAlgorithm
 from causy.serialization import serialize_algorithm, load_algorithm_from_specification
 
 from tests.utils import CausyTestCase
